+++ Tests/GRDBTests/ValueObservationMapTests.swift
@@ -33,24 +33,24 @@ class ValueObservationMapTests: GRDBTestCase {
             })
             
             // Create an observation
-            var observation = ValueObservation
+            let observation = ValueObservation
                 .tracking(DatabaseRegion.fullDatabase, reducer: { _ in reducer })
                 .map { count -> String in return "\(count)" }
-            observation.extent = .databaseLifetime
             
             // Start observation
-            _ = try observation.start(in: dbWriter) { count in
+            let observer = try observation.start(in: dbWriter) { count in
                 counts.append(count)
                 notificationExpectation.fulfill()
             }
-            
-            try dbWriter.writeWithoutTransaction { db in
-                try db.execute(sql: "INSERT INTO t DEFAULT VALUES")
-                try db.execute(sql: "INSERT INTO t DEFAULT VALUES")
+            try withExtendedLifetime(observer) {
+                try dbWriter.writeWithoutTransaction { db in
+                    try db.execute(sql: "INSERT INTO t DEFAULT VALUES")
+                    try db.execute(sql: "INSERT INTO t DEFAULT VALUES")
+                }
+                
+                waitForExpectations(timeout: 1, handler: nil)
+                XCTAssertEqual(counts, ["1", "2", "3"])
             }
-            
-            waitForExpectations(timeout: 1, handler: nil)
-            XCTAssertEqual(counts, ["1", "2", "3"])
         }
         
         try test(makeDatabaseQueue())
@@ -59,12 +59,10 @@ class ValueObservationMapTests: GRDBTestCase {
     
     func testMapPreservesConfiguration() {
         var observation = ValueObservation.tracking(DatabaseRegion(), fetch: { _ in })
-        observation.extent = .nextTransaction
         observation.requiresWriteAccess = true
         observation.scheduling = .unsafe(startImmediately: true)
         
         let mappedObservation = observation.map { _ in }
-        XCTAssertEqual(mappedObservation.extent, observation.extent)
         XCTAssertEqual(mappedObservation.requiresWriteAccess, observation.requiresWriteAccess)
         switch mappedObservation.scheduling {
         case .unsafe: